pyproject.toml
[project]
name = "napari-deeplabcut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dask-image>=2024.5.3",
    "matplotlib>=3.9.2",
    "napari==0.4.19",
    "natsort>=8.4.0",
    "numpy>=1.21",
    "opencv-python-headless>=4.10.0.84",
    "pandas>=2.2.3",
    "pyside6==6.7.3",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-qt>=4.4.0",
    "pyyaml>=6.0.2",
    "qtpy>=2.4.1",
    "scikit-image>=0.24.0",
    "scipy>=1.13.1",
    "tables>=3.9.2",
    "torch >= 2.4.1",
    "tox>=4.20.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points."napari.manifest"]
"napari-deeplabcut" = "napari_deeplabcut:napari.yaml"

[tool.hatch.metadata]
allow-direct-references = true